factor out toolbar matchers
