--- lib/Editor/EditorToolbarArea.tsx
@@ -1,231 +1,0 @@
-import React, { useEffect, useRef, useState } from "react"
-import { Editor, Path, Range } from "slate"
-import type { DOMPoint } from "slate-dom"
-import type { HistoryEditor } from "slate-history"
-import { ReactEditor, useSlate, useSlateSelection } from "slate-react"
-import { FormattingToolbarContent } from "./FormattingToolbarContent"
-import { LinkToolbarContent } from "./LinkToolbarContent"
-import { isLinkElement, type LinkElement as LinkElementNode } from "./types"
-import { ToolbarArea } from "~/Components/ToolbarArea"
-import { isFormattableRange } from "~/helpers/range"
-
-export type ToolbarContentProps = {
-  pinOpen: () => void
-  unpinOpen: () => void
-}
-
-function domPointFromClientXY(x: number, y: number): DOMPoint | null {
-  const doc = document
-  const pos = doc.caretPositionFromPoint?.(x, y)
-  if (!pos) return null
-  return [pos.offsetNode, pos.offset]
-}
-
-function pathsEqual(a: Path | null, b: Path | null): boolean {
-  if (a === null && b === null) return true
-  if (a === null || b === null) return false
-  return Path.equals(a, b)
-}
-
-type SlateEditor = ReactEditor & HistoryEditor
-
-function linkPathAtPoint(
-  editor: SlateEditor,
-  point: { path: Path; offset: number }
-): Path | null {
-  const above = Editor.above(editor, {
-    at: point,
-    match: isLinkElement,
-  })
-  return above ? above[1] : null
-}
-
-function isValidLinkPath(editor: SlateEditor, path: Path): boolean {
-  try {
-    const [n] = Editor.node(editor, path)
-    return isLinkElement(n)
-  } catch {
-    return false
-  }
-}
-
-type EditorToolbarAreaProps = {
-  ghostSelection: Range | undefined
-  children: React.ReactNode
-}
-
-export const EditorToolbarArea: React.FC<EditorToolbarAreaProps> = ({
-  ghostSelection,
-  children,
-}) => {
-  const toolbarListenerAreaRef = useRef<HTMLDivElement>(null)
-  const editor = useSlate() as SlateEditor
-  const selection = useSlateSelection()
-
-  const [hoverPath, setHoverPath] = useState<Path | null>(null)
-  const hoverPathRef = useRef<Path | null>(null)
-  const [pinnedPath, setPinnedPath] = useState<Path | null>(null)
-
-  const focused = ReactEditor.isFocused(editor)
-
-  const pathAtCaret =
-    selection == null || !focused || !Range.isCollapsed(selection)
-      ? null
-      : linkPathAtPoint(editor, selection.anchor)
-
-  const pinStillValid =
-    pinnedPath !== null && isValidLinkPath(editor, pinnedPath)
-
-  useEffect(() => {
-    if (pinnedPath !== null && !pinStillValid) {
-      setPinnedPath(null)
-    }
-  }, [pinnedPath, pinStillValid])
-
-  const activePath = (() => {
-    if (pinnedPath && isValidLinkPath(editor, pinnedPath)) {
-      return pinnedPath
-    }
-    if (focused && pathAtCaret) return pathAtCaret
-    return hoverPath
-  })()
-
-  // Formatting toolbar: expanded selection while focused, else ghost selection
-  // when blurred; suppressed in code/frozen blocks; needs a non-empty DOM rect.
-  const hasExpandedSelection = selection && !Range.isCollapsed(selection)
-  const formattingCandidateRange = hasExpandedSelection
-    ? selection
-    : focused
-    ? undefined
-    : ghostSelection
-
-  let formattingTargetRect: DOMRect | undefined
-  if (isFormattableRange(editor, formattingCandidateRange)) {
-    try {
-      const domRange = ReactEditor.toDOMRange(editor, formattingCandidateRange)
-      const rect = domRange.getClientRects()[0]
-      if (rect && !(rect.width === 0 && rect.height === 0)) {
-        formattingTargetRect = rect
-      }
-    } catch {
-      // toDOMRange can throw when Slate's range no longer maps cleanly to DOM
-    }
-  }
-
-  const activeRange =
-    formattingTargetRect && formattingCandidateRange
-      ? formattingCandidateRange
-      : null
-  const targetRect = activeRange ? formattingTargetRect : undefined
-  const showFormatting = Boolean(activeRange && targetRect)
-
-  useEffect(() => {
-    const area = toolbarListenerAreaRef.current
-    if (!area) return
-
-    const setHoverIfChanged = (next: Path | null) => {
-      if (pathsEqual(hoverPathRef.current, next)) return
-      hoverPathRef.current = next
-      setHoverPath(next)
-    }
-
-    const onPointerMove = (e: PointerEvent) => {
-      const el = document.elementFromPoint(e.clientX, e.clientY)
-      if (!el || !area.contains(el)) {
-        setHoverIfChanged(null)
-        return
-      }
-      const domPoint = domPointFromClientXY(e.clientX, e.clientY)
-      if (!domPoint || !area.contains(domPoint[0])) {
-        setHoverIfChanged(null)
-        return
-      }
-      const slatePoint = ReactEditor.toSlatePoint(editor, domPoint, {
-        exactMatch: false,
-        suppressThrow: true,
-      })
-      if (!slatePoint) {
-        setHoverIfChanged(null)
-        return
-      }
-      const nextPath = linkPathAtPoint(editor, slatePoint)
-      setHoverIfChanged(nextPath)
-    }
-
-    const onPointerLeave = () => {
-      setHoverIfChanged(null)
-    }
-
-    area.addEventListener("pointermove", onPointerMove)
-    area.addEventListener("pointerleave", onPointerLeave)
-    return () => {
-      area.removeEventListener("pointermove", onPointerMove)
-      area.removeEventListener("pointerleave", onPointerLeave)
-    }
-  }, [editor])
-
-  const pinOpen = () => {
-    if (activePath) setPinnedPath(activePath)
-  }
-
-  const unpinOpen = () => {
-    setPinnedPath(null)
-  }
-
-  let toolbar:
-    | { target: Element | DOMRectReadOnly; content: React.ReactNode }
-    | undefined
-
-  if (showFormatting && activeRange && targetRect) {
-    toolbar = {
-      target: targetRect,
-      content: <FormattingToolbarContent activeRange={activeRange} />,
-    }
-  } else if (activePath && isValidLinkPath(editor, activePath)) {
-    let linkNode: LinkElementNode
-    try {
-      const [n] = Editor.node(editor, activePath)
-      if (!isLinkElement(n)) return null
-      linkNode = n
-    } catch {
-      return null
-    }
-
-    let linkDom: HTMLElement
-    try {
-      linkDom = ReactEditor.toDOMNode(editor, linkNode)
-    } catch {
-      return null
-    }
-
-    toolbar = {
-      target: linkDom,
-      content: (
-        <>
-          <LinkToolbarContent
-            key={JSON.stringify(activePath)}
-            linkPath={activePath}
-            linkNode={linkNode}
-            pinOpen={pinOpen}
-            unpinOpen={unpinOpen}
-          />
-        </>
-      ),
-    }
-  }
-
-  if (!toolbar) {
-    return <>{children}</>
-  }
-
-  return (
-    <ToolbarArea
-      target={toolbar.target}
-      open
-      ref={toolbarListenerAreaRef}
-      content={toolbar.content}
-    >
-      {children}
-    </ToolbarArea>
-  )
-}
